app/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
pandas==3.0.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6

# Database
SQLAlchemy==2.0.36
PyMySQL==1.1.1
cryptography==44.0.0

# Environment management
python-dotenv==1.0.1

# Authentication (Auth0)
python-jose[cryptography]==3.3.0
Flask-Session==0.8.0
